Added authentication

+++ base/views.py
@@ -1,10 +1,62 @@
 from django.shortcuts import render, redirect
 from .models import Room,Topic
+from django.contrib.auth.decorators import login_required
 from .forms import RoomForm
 from django.db.models import Q
-# Create your views here.
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib import messages
+
+from django.http import HttpRequest, HttpResponse
+
+def loginPage(request):
+    page= 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username= request.POST.get('username').lower()
+        password= request.POST.get('password')
+
+        try:
+            user= User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user= authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username OR password does not exist')
 
 
+    context={'page':page}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    
+    form = UserCreationForm()
+    if request.method=='POST':
+        form= UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit= False)
+            user.username= user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error occured during registeration')
+    
+
+    return render(request, 'base/login_register.html',{'form': form})
 
 def home(request):
     q= request.GET.get('q')  if request.GET.get('q')!= None else ''
@@ -25,6 +77,7 @@ def room(request,pk):
 
     return render(request,'base/room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method== 'POST':
@@ -35,9 +88,14 @@ def createRoom(request):
     context ={'form': form}
     return render(request, 'base/room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room= Room.objects.get(id=pk)
     form= RoomForm(instance=room)
+
+    if request.user != room.host :
+        return HttpResponse('You are not allowed here! ')
+
     if request.method== 'POST':
         form=RoomForm(request.POST,instance= room)
         if form.is_valid():
@@ -46,8 +104,13 @@ def updateRoom(request,pk):
     context={'form':form}
     return render(request, 'base/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
+
     room=Room.objects.get(id=pk)
+    if request.user != room.host :
+        return HttpResponse('You are not allowed here! ')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
